Add changes to table
