+++ src/views/domain-workflows/domain-workflows-table/domain-workflows-table.styles.ts
@@ -4,10 +4,4 @@ export const styled = {
   TableContainer: createStyled('div', {
     overflowX: 'auto',
   }),
-  ErrorPanelContainer: createStyled('div', ({ $theme }) => ({
-    padding: `${$theme.sizing.scale1200} 0px`,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  })),
 };
